Move current transaction related functions to the transaction module in transaction.rs and re-export functions in cli.rs
